+++ config/dev.exs
@@ -1,4 +1,6 @@
 import Config
 
 config :elixir_lokalise_api,
-  api_token: {:system, "LOKALISE_API_TOKEN"}
+  api_token: {:system, "LOKALISE_API_TOKEN"},
+  oauth2_client_id: {:system, "OAUTH2_CLIENT_ID"},
+  oauth2_client_secret: {:system, "OAUTH2_CLIENT_SECRET"}